project2/codes/requirements.txt
accelerate==1.11.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.1
bitsandbytes==0.48.2
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
dash==3.3.0
dash_cytoscape==1.0.2
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.20.0
Flask==3.1.2
fonttools==4.60.1
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mpmath==1.3.0
narwhals==2.12.0
neo4j==6.0.3
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==2.8.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
plotly==6.5.0
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
retrying==1.4.2
safetensors==0.6.2
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
zipp==3.23.0
